big update

+++ booksItems.php
@@ -1,22 +1,18 @@
 <?php
 defined('CORE_INDEX') or die('restricted access');
 
-	$fileData = file_get_contents('data.json');
-	$dataJson = json_decode($fileData);
-	//var_dump($dataJson->books[0]->title) ;
-	//$dataJson->books
-	$books = $dataJson->books;
-
 ?>
 <h1>Книжный фонд</h1>
 
-<a class="add" href="index.php?action=formBook" onclick="add">[+]</a>
-
+<?php if ($canEdit): ?>
+	<a class="add" href="index.php?action=bookForm" onclick="add">[+]</a>
+<?php endif; ?>
 <table>
 	<tr>
 		<th>Наименование</th>
 		<th>Описание</th>
 		<th>Авторы</th>
+		<th></th>
 	</tr>
 	<tbody>
 		<?php foreach ($books as $key => $value): ?>
@@ -25,8 +21,10 @@ defined('CORE_INDEX') or die('restricted access');
 				<td><?= $books[$key]->description ?></td>
 				<td><?= $books[$key]->author ?></td>
 				<td>
-					<a class="edit" href="index.php?action=formBook&idBook=<?= $books[$key]->id ?>">[<->]</a>
-					<a class="delete" href="index.php?action=deleteBook&idBook=<?= $books[$key]->id ?>">[-]</a>
+					<?php if ($canEdit): ?>
+						<a class="edit" href="index.php?action=bookForm&idBook=<?= $books[$key]->id ?>">[<->]</a>
+						<a class="delete" href="index.php?action=deleteBook&idBook=<?= $books[$key]->id ?>">[-]</a>
+					<?php endif; ?>
 				</td>
 			</tr>
 		<?php endforeach; ?>
